Trait colored_truecolor::Colorize

pub trait Colorize {

    fn black(self) -> ColoredString;
    fn red(self) -> ColoredString;
    fn green(self) -> ColoredString;
    fn yellow(self) -> ColoredString;
    fn blue(self) -> ColoredString;
    fn magenta(self) -> ColoredString;
    fn purple(self) -> ColoredString;
    fn cyan(self) -> ColoredString;
    fn white(self) -> ColoredString;
    fn bright_black(self) -> ColoredString;
    fn bright_red(self) -> ColoredString;
    fn bright_green(self) -> ColoredString;
    fn bright_yellow(self) -> ColoredString;
    fn bright_blue(self) -> ColoredString;
    fn bright_magenta(self) -> ColoredString;
    fn bright_purple(self) -> ColoredString;
    fn bright_cyan(self) -> ColoredString;
    fn bright_white(self) -> ColoredString;
    fn palette(self, code: u8) -> ColoredString;
    fn true_color(self, r: u8, g: u8, b: u8) -> ColoredString;
    fn hex_color(self, color: u64) -> ColoredString;
    fn color<S: Into<Color>>(self, color: S) -> ColoredString;
    fn on_black(self) -> ColoredString;
    fn on_red(self) -> ColoredString;
    fn on_green(self) -> ColoredString;
    fn on_yellow(self) -> ColoredString;
    fn on_blue(self) -> ColoredString;
    fn on_magenta(self) -> ColoredString;
    fn on_purple(self) -> ColoredString;
    fn on_cyan(self) -> ColoredString;
    fn on_white(self) -> ColoredString;
    fn on_bright_black(self) -> ColoredString;
    fn on_bright_red(self) -> ColoredString;
    fn on_bright_green(self) -> ColoredString;
    fn on_bright_yellow(self) -> ColoredString;
    fn on_bright_blue(self) -> ColoredString;
    fn on_bright_magenta(self) -> ColoredString;
    fn on_bright_purple(self) -> ColoredString;
    fn on_bright_cyan(self) -> ColoredString;
    fn on_bright_white(self) -> ColoredString;
    fn on_palette(self, code: u8) -> ColoredString;
    fn on_true_color(self, r: u8, g: u8, b: u8) -> ColoredString;
    fn on_hex_color(self, color: u64) -> ColoredString;
    fn on_color<S: Into<Color>>(self, color: S) -> ColoredString;
    fn clear(self) -> ColoredString;
    fn normal(self) -> ColoredString;
    fn bold(self) -> ColoredString;
    fn dimmed(self) -> ColoredString;
    fn italic(self) -> ColoredString;
    fn underline(self) -> ColoredString;
    fn blink(self) -> ColoredString;
    fn reverse(self) -> ColoredString;
    fn reversed(self) -> ColoredString;
    fn hidden(self) -> ColoredString;
    fn strikethrough(self) -> ColoredString;

}

The trait that enables something to be given color.

You can use colored effectively simply by importing this trait and then using its methods on String and &str.

Required Methods

fn black(self) -> ColoredString

fn red(self) -> ColoredString

fn green(self) -> ColoredString

fn yellow(self) -> ColoredString

fn blue(self) -> ColoredString

fn magenta(self) -> ColoredString

fn purple(self) -> ColoredString

fn cyan(self) -> ColoredString

fn white(self) -> ColoredString

fn bright_black(self) -> ColoredString

fn bright_red(self) -> ColoredString

fn bright_green(self) -> ColoredString

fn bright_yellow(self) -> ColoredString

fn bright_blue(self) -> ColoredString

fn bright_magenta(self) -> ColoredString

fn bright_purple(self) -> ColoredString

fn bright_cyan(self) -> ColoredString

fn bright_white(self) -> ColoredString

fn palette(self, code: u8) -> ColoredString

fn true_color(self, r: u8, g: u8, b: u8) -> ColoredString

fn hex_color(self, color: u64) -> ColoredString

fn color<S: Into<Color>>(self, color: S) -> ColoredString

fn on_black(self) -> ColoredString

fn on_red(self) -> ColoredString

fn on_green(self) -> ColoredString

fn on_yellow(self) -> ColoredString

fn on_blue(self) -> ColoredString

fn on_magenta(self) -> ColoredString

fn on_purple(self) -> ColoredString

fn on_cyan(self) -> ColoredString

fn on_white(self) -> ColoredString

fn on_bright_black(self) -> ColoredString

fn on_bright_red(self) -> ColoredString

fn on_bright_green(self) -> ColoredString

fn on_bright_yellow(self) -> ColoredString

fn on_bright_blue(self) -> ColoredString

fn on_bright_magenta(self) -> ColoredString

fn on_bright_purple(self) -> ColoredString

fn on_bright_cyan(self) -> ColoredString

fn on_bright_white(self) -> ColoredString

fn on_palette(self, code: u8) -> ColoredString

fn on_true_color(self, r: u8, g: u8, b: u8) -> ColoredString

fn on_hex_color(self, color: u64) -> ColoredString

fn on_color<S: Into<Color>>(self, color: S) -> ColoredString

fn clear(self) -> ColoredString

fn normal(self) -> ColoredString

fn bold(self) -> ColoredString

fn dimmed(self) -> ColoredString

fn italic(self) -> ColoredString

fn underline(self) -> ColoredString

fn blink(self) -> ColoredString

fn reverse(self) -> ColoredString

Historical name of Colorize::reversed. May be removed in a future version. Please use Colorize::reversed instead

fn reversed(self) -> ColoredString

This should be preferred to Colorize::reverse.

fn hidden(self) -> ColoredString

fn strikethrough(self) -> ColoredString